check/out use a content-aware sha strategy
This adds BoshDeploymentResource::BoshManifest#shasum.

It will load the manifest, sort any hashes, and then add to the digest
each key and its values. It will not sort arrays (they are presumed to
not change order due to marshaling/unmarshaling of the manifest YAML).

This avoids the version returned by check changing due to whitespace or
sorting of the hashes.

[#115707613]